ci: add GitHub Actions for CI/CD process
This commit introduces a build pipeline via GitHub Actions to improve our continuous integration and deployment(CI/CD). It is launched for changes pushed to the master, bugfix, hotfix, or release branches excluding markdown file modifications and pull-request events. It consists of several stages such as, the initial commit check, the Node.js environment configuration, and finally, the changelog generation and the creation of a draft release on GitHub. This improves developer productivity by standardizing and automating build, test, and deployment processes.