ci: switch package manager from pnpm to yarn temporary
Changes have been made to switch the package manager from pnpm to Yarn in the build workflow. As yarn has better performance and speed compared to pnpm, it has been utilized to install dependencies, build the application, and execute tests. This change will contribute to increased efficiency in the project's build process.